Extract ClassSourceTests from GenericSourceTests
